merges/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-bigquery
google-cloud-documentai
python-dotenv

